Added ability to fetch all DB changes and updates.

+++ BungieNetCoreAPI/Authrorization/AuthorizationTokenData.cs
@@ -10,11 +10,11 @@ namespace NetBungieAPI.Authrorization
         public int ExpiresIn { get; }
         public string RefreshToken { get; }
         public int RefreshExpiresIn { get; }
-        public int MembershipId { get; }
+        public long MembershipId { get; }
         public DateTime ReceiveTime { get; } = DateTime.Now;
 
         [JsonConstructor]
-        internal AuthorizationTokenData(string access_token, string token_type, int expires_in, string refresh_token, int refresh_expires_in, int membership_id)
+        internal AuthorizationTokenData(string access_token, string token_type, int expires_in, string refresh_token, int refresh_expires_in, long membership_id)
         {
             AccessToken = access_token;
             TokenType = token_type;
